+++ diary-backend/README.md
@@ -1,1 +1,3 @@
 # Diary Backend Service
+## Entities
+* 회원: io.doubleu0714.diary.entity.Member
