+++ diary-backend/src/main/java/io/doubleu0714/diary/entity/BaseInfo.java
@@ -1,4 +1,4 @@
-package io.doubleu0714.diary.diarybackend.entity;
+package io.doubleu0714.diary.entity;
 
 import java.time.LocalDateTime;
